backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
litellm
requests
nltk
haystack-ai
markdown-it-py
mdit_plain
pinecone_haystack
scipy
langchain
langchain-core
langgraph
snowflake-connector-python[pandas]